патч по замечаниям
