+++ .gitignore
@@ -2,6 +2,7 @@
 .gradle
 /local.properties
 /.idea/caches
+/.idea
 /.idea/libraries
 /.idea/modules.xml
 /.idea/workspace.xml
