+++ app/src/main/java/acr/appcradle/shoppinglist/ui/components/AppInputFields.kt
@@ -1,0 +1,62 @@
+package acr.appcradle.shoppinglist.ui.components
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Clear
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.unit.dp
+
+object AppInputFields {
+    @Composable
+    fun MainInputField(
+        modifier: Modifier = Modifier,
+        placeholderText: String,
+        isSearchIconNeeded: Boolean = false,
+        onValueChange: (String) -> Unit = {}
+    ) {
+        var inputText by rememberSaveable { mutableStateOf("") }
+        TextField(
+            modifier = modifier,
+            value = inputText,
+            placeholder = { Text(text = placeholderText) },
+            onValueChange = {
+                inputText = it
+                onValueChange(inputText)
+            },
+            colors = TextFieldDefaults.colors(
+                focusedIndicatorColor = Color.Transparent,
+                unfocusedIndicatorColor = Color.Transparent
+            ),
+            shape = RoundedCornerShape(10.dp),
+            singleLine = true,
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
+            trailingIcon = {
+                if (inputText != "")
+                    Icon(
+                        modifier = Modifier.clickable { inputText = "" },
+                        imageVector = Icons.Default.Clear, contentDescription = null
+                    )
+            },
+            leadingIcon = {
+                if (isSearchIconNeeded)
+                    Icon(
+                        imageVector = Icons.Default.Search, contentDescription = null
+                    )
+            }
+        )
+    }
+}
