+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/list_creation/ListCreationScreen.kt
@@ -1,15 +1,11 @@
 package acr.appcradle.shoppinglist.ui.screens.list_creation
 
-import acr.appcradle.shoppinglist.RoutesList
 import acr.appcradle.shoppinglist.model.IconsIntent
-import acr.appcradle.shoppinglist.model.ListRepository
 import acr.appcradle.shoppinglist.model.NewListData
 import acr.appcradle.shoppinglist.ui.AppViewModel
-import acr.appcradle.shoppinglist.ui.components.AppLargeButton
+import acr.appcradle.shoppinglist.ui.components.AppInputFields
 import acr.appcradle.shoppinglist.ui.components.AppNavTopBar
-import acr.appcradle.shoppinglist.ui.components.appInputField
-import acr.appcradle.shoppinglist.ui.theme.ShoppingListTheme
-import acr.appcradle.shoppinglist.utils.ThemePreviews
+import acr.appcradle.shoppinglist.ui.components.ShoppingListButtons
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -19,11 +15,10 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -59,7 +54,7 @@ fun ListCreationScreenUi(
     onNextClick: () -> Unit
 ) {
     val scroll = rememberScrollState()
-    var inputText by remember { mutableStateOf("") }
+    var inputText by rememberSaveable { mutableStateOf("") }
     val viewModel: AppViewModel = hiltViewModel()
 
     Scaffold(
@@ -68,7 +63,6 @@ fun ListCreationScreenUi(
                 isBackIconEnable = true,
                 title = "Создать список",
                 onBackIconClick = { onBackClick() },
-                screenRoute = RoutesList.ListCreationRoute
             )
         }
     ) { innerPaddings ->
@@ -79,11 +73,12 @@ fun ListCreationScreenUi(
                 .verticalScroll(scroll),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            inputText = appInputField(
+            AppInputFields.MainInputField(
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(horizontal = 16.dp),
-                placeholderText = "Введите название списка"
+                placeholderText = "Введите название списка",
+                onValueChange = { inputText = it }
             )
             ColorPalette(
                 modifier = Modifier.padding(vertical = 16.dp),
@@ -95,13 +90,16 @@ fun ListCreationScreenUi(
                 onIconClick = onIconClick
             )
             Spacer(Modifier.weight(1f))
-            AppLargeButton(onClick = {
-                if (inputText.isNotBlank()) {
-                    viewModel.createNewList(inputText.trim()) {
-                        onNextClick()
+            ShoppingListButtons.AppLargeButton(
+                onClick = {
+                    if (inputText.isNotBlank()) {
+                        viewModel.createNewList(inputText.trim()) {
+                            onNextClick()
+                        }
                     }
-                }
-            }, text = "Создать")
+                },
+                text = "Создать"
+            )
         }
     }
 }
