+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/list_creation/ListOfColors.kt
@@ -1,0 +1,12 @@
+package acr.appcradle.shoppinglist.ui.screens.list_creation
+
+import acr.appcradle.shoppinglist.ui.theme.Team2Colors
+import androidx.compose.ui.graphics.Color
+
+val listOfColors = listOf<Color>(
+    Team2Colors.additionalColorGreen,
+    Team2Colors.additionalColorPurple,
+    Team2Colors.additionalColorBlue,
+    Team2Colors.additionalColorRed,
+    Team2Colors.additionalColorYellow
+)
