+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/list_creation/ListOfIcons.kt
@@ -1,0 +1,26 @@
+package acr.appcradle.shoppinglist.ui.screens.list_creation
+
+import acr.appcradle.shoppinglist.R
+
+val listOfIcons = listOf<Int>(
+    R.drawable.icon_snowflake,
+    R.drawable.airplane_outline,
+    R.drawable.alert_outline,
+    R.drawable.balloon_outline,
+    R.drawable.bandage_outline,
+    R.drawable.barbell_outline,
+    R.drawable.fast_food_outline,
+    R.drawable.home_outline,
+    R.drawable.library_outline,
+    R.drawable.wine_outline,
+    R.drawable.bed_outline,
+    R.drawable.briefcase_outline,
+    R.drawable.build_outline,
+    R.drawable.business_outline,
+    R.drawable.calendar_number_outline,
+    R.drawable.color_palette_outline,
+    R.drawable.cart_outline,
+    R.drawable.car_outline,
+    R.drawable.gift_outline,
+    R.drawable.paw_outline,
+)
