+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/lists_all/EmptyListAllUi.kt
@@ -1,7 +1,7 @@
 package acr.appcradle.shoppinglist.ui.screens.lists_all
 
 import acr.appcradle.shoppinglist.R
-import acr.appcradle.shoppinglist.ui.components.AppLargeButton
+import acr.appcradle.shoppinglist.ui.components.ShoppingListButtons
 import acr.appcradle.shoppinglist.ui.theme.Typography
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
@@ -21,28 +21,22 @@ import androidx.compose.ui.unit.dp
 fun EmptyListAllUi(onNextClick: () -> Unit) {
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
         Spacer(Modifier.height(104.dp))
-
         Image(
             painterResource(R.drawable.empty_list_image),
             contentDescription = null
         )
-
         Spacer(Modifier.height(52.dp))
-
         Text(
             text = "Давайте спланируем покупки!",
             style = Typography.titleLarge.copy(textAlign = TextAlign.Center),
         )
-
         Text(
             modifier = Modifier.padding(top = 12.dp),
             text = "Создайте свой первый список",
             style = Typography.bodyLarge.copy(textAlign = TextAlign.Center)
         )
-
         Spacer(Modifier.weight(1f))
-
-        AppLargeButton(
+        ShoppingListButtons.AppLargeButton(
             text = "Создать список",
             onClick = {
                 onNextClick()
@@ -54,5 +48,5 @@ fun EmptyListAllUi(onNextClick: () -> Unit) {
 @Preview
 @Composable
 private fun EmptyListPreview() {
-    EmptyListAllUi() {}
+    EmptyListAllUi {}
 }
