+++ app/src/main/java/acr/appcradle/shoppinglist/ui/theme/Color.kt
@@ -12,12 +12,15 @@ val Pink40 = Color(0xFF7D5260)
 
 
 //Custom
-val tortoise = Color(0xFF147A85)
-val iconsBg = Color(0xFFCBCCCC)
-val gray = Color(0xFF74797E)
+object Team2Colors {
+    val tortoise = Color(0xFF147A85)
+    val iconsBg = Color(0xFFCBCCCC)
+    val team2colors_gray = Color(0xFF74797E)
 
-val additionalColorGreen = Color(0xFFB4D8CC)
-val additionalColorPurple = Color(0xFFBDBAD8)
-val additionalColorBlue = Color(0xFF96CDE4)
-val additionalColorRed = Color(0xFFE7ABA7)
-val additionalColorYellow = Color(0xFFFCE7A2)
+    val additionalColorGreen = Color(0xFFB4D8CC)
+    val additionalColorPurple = Color(0xFFBDBAD8)
+    val additionalColorBlue = Color(0xFF96CDE4)
+    val additionalColorRed = Color(0xFFE7ABA7)
+    val additionalColorYellow = Color(0xFFFCE7A2)
+
+}
